make quality-scaler command plus lazy ai model download. pip install . && quality-scaler
