+++ requirements.testing.txt
@@ -1,0 +1,9 @@
+black
+isort
+flake8
+mypy
+pylint
+pytest
+tox
+ruff
+pytest-xdist
