--- requirements.txt
@@ -1,11 +1,0 @@
-#AI
-torch-directml==0.1.13.1.dev230413
-
-#GUI
-customtkinter
-packaging
-
-#UTILS
-pyinstaller==5.13.2
-moviepy
-opencv-python-headless
