+++ src/quality_scaler/AI/readme.txt
@@ -1,0 +1,1 @@
+Files will be downloaded here.
