+++ tests/test_cli.py
@@ -1,0 +1,20 @@
+"""
+Unit test file.
+"""
+import os
+import unittest
+
+COMMAND = "quality_scaler"
+
+
+class MainTester(unittest.TestCase):
+    """Main tester class."""
+
+    def test_imports(self) -> None:
+        """Test command line interface (CLI)."""
+        rtn = os.system(COMMAND)
+        self.assertEqual(0, rtn)
+
+
+if __name__ == "__main__":
+    unittest.main()
