+++ tox.ini
@@ -1,0 +1,24 @@
+# content of: tox.ini , put in same dir as setup.py
+[tox]
+envlist = py310
+
+[gh-actions]
+python =
+    3.10: py310
+
+[flake8]
+per-file-ignores = __init__.py:F401
+ignore = E501, E203, W503
+
+[testenv]
+# install pytest in the virtualenv where commands will be executed
+deps =
+    -r{toxinidir}/requirements.testing.txt
+commands =
+    # NOTE: you can run any command line tool here - not just tests
+    ruff src
+    ruff tests
+    flake8 src tests
+    pylint src tests
+    mypy src tests
+    pytest tests -n auto -vv
